:zap: :: hooks로 분리
